Update to new release model
Alas, Maven Repository Plugin is disappointing: when used in the simple
way with Maven GPG Plugin as we did, it repackages some artifacts after
signing them, which breaks the signatures.

There is probably no easy solution and using Maven Release Plugin is a
big burden at this moment: seriously, spending more time on dealing with
these annoyances than with actual code development :-(

Simple solution: make all as local as possible and make the release
bundles with a script. It is not so elegant, definitely not as automated,
but it work without much fuss and it could help to think twice before
releasing a thing.